1760. Minimum Limit of Balls in a Bag

+++ Important Questions/Minimum_Limit_of_Balls_Bag.java
@@ -3,7 +3,7 @@ import java.util.*;
 public class Minimum_Limit_of_Balls_Bag {
 
     public static boolean isPossible(int mid, int nums[], int maxOps) {
-        int total = 0;
+        long total = 0;
 
         for (int num : nums) {
             int ops = num / mid;
